feat: adds body parser
