+++ lesson_24/public/contact.html
@@ -26,7 +26,7 @@
 
     <div class="contact-form">
         <h2>Contact Us</h2>
-        <form action="/results.html" method="POST">
+        <form action="/contact" method="POST">
             <label for="first-name">Full Name</label>
             <div class="input-group">
                 <input type="text" id="first-name" name="first-name" placeholder="First Name" required>
